feat: add `telefunc/universal-middleware` and `telefunc/vike` (#381)
Add a Telefunc universal middleware exported at `telefunc/universal-middleware`
and a Vike extension exported at `telefunc/vike` that automatically adds
Telefunc's universal middleware to Vike's `+middleware`.

- `telefunc/universal-middleware` — built on the streaming `serve()` API and the
  universal-middleware standard, so it works with any server (Hono, Express,
  Cloudflare, Vercel, ...). Its default export is the ready-to-use enhanced
  middleware; the named `telefuncMiddleware` export is a factory (e.g. for a
  custom `telefuncUrl`).
- `telefunc/vike` — a Vike extension config that provides `middleware` via a
  pointer import (`import:telefunc/universal-middleware:default`) so Vike imports
  it server-side only. It's built to `+config.js` so Vike recognizes it as an
  extension config (the convention for packages not named `vike-*`).

Usage:

```ts
// +config.ts — zero-boilerplate via the Vike extension
import telefunc from 'telefunc/vike'
export default { extends: [telefunc] }
```

```ts
// +middleware.ts — or wire the middleware up manually
export { default } from 'telefunc/universal-middleware'
```

`examples/vike` now uses the `telefunc/vike` extension, dropping its
`+middleware.ts` boilerplate.
