--- examples/vike/+middleware.ts
@@ -1,26 +1,0 @@
-export { middlewareTelefunc as default }
-
-import { enhance, type UniversalMiddleware } from '@universal-middleware/core'
-import { serve } from 'telefunc'
-
-const telefuncUniversalMiddleware: UniversalMiddleware = async (request, context, runtime) => {
-  const httpResponse = await serve({
-    request,
-    context: {
-      ...context,
-      ...runtime,
-    },
-  })
-  const { statusCode, headers } = httpResponse
-
-  return new Response(httpResponse.getReadableWebStream(), {
-    status: statusCode,
-    headers,
-  })
-}
-
-const middlewareTelefunc = enhance(telefuncUniversalMiddleware, {
-  name: 'telefunc',
-  method: 'POST',
-  path: '/_telefunc',
-})
