+++ examples/vike/pages/+config.ts
@@ -1,6 +1,7 @@
 import vikeReact from 'vike-react/config'
+import telefunc from 'telefunc/vike'
 import type { Config } from 'vike/types'
 
 export default {
-  extends: [vikeReact],
+  extends: [vikeReact, telefunc],
 } satisfies Config
