+++ packages/telefunc/client/poisen-pills/universal-middleware.ts
@@ -1,0 +1,1 @@
+throw new Error('Module telefunc/universal-middleware cannot be loaded on the client-side')
