+++ packages/telefunc/client/poisen-pills/vike.ts
@@ -1,0 +1,1 @@
+throw new Error('Module telefunc/vike cannot be loaded on the client-side')
