+++ packages/telefunc/universal-middleware.d.ts
@@ -1,0 +1,3 @@
+// Help TS's resolver until it supports `package.json#exports`
+export * from './dist/node/universal-middleware/index.js'
+export { default } from './dist/node/universal-middleware/index.js'
