+++ packages/telefunc/universal-middleware.js
@@ -1,0 +1,3 @@
+// Some tools don't support `package.json#exports`
+export * from './dist/node/universal-middleware/index.js'
+export { default } from './dist/node/universal-middleware/index.js'
