+++ packages/telefunc/vike.d.ts
@@ -1,0 +1,2 @@
+// Help TS's resolver until it supports `package.json#exports`
+export { default } from './dist/node/vike/+config.js'
