+++ packages/telefunc/vike.js
@@ -1,0 +1,2 @@
+// Some tools don't support `package.json#exports`
+export { default } from './dist/node/vike/+config.js'
